BUILD(client): Cross-compile the Windows client with MinGW
Adds a containerised MinGW-w64 cross-build so the Windows client can be produced
from Linux, and fixes what that build found.

Most of the source changes are in code that only compiles on Windows and had
therefore only ever been compiled by MSVC. GCC's warning set is different, and
this project builds with -Werror, so a large number of narrowing conversions,
function-pointer casts through FARPROC, and printf format mismatches surfaced at
once. The casts are all written to preserve the existing behaviour exactly; where
the old code truncated a value, it still truncates, and the cast just says so.

Two portability bugs were real rather than cosmetic:

  * The positional-audio plugins selected their backend on WIN32, which the MSVC
    project templates define but the compiler does not, so a cross-build fell
    through to the Linux backend and failed on <elf.h>.
  * win.h declared QOS_FLOWID itself to work around a gap in MinGW's <qos2.h>.
    Current mingw-w64 declares it, as ULONG rather than the UINT32 assumed here,
    so the workaround had become a conflicting definition.

The client's CMake also assumed MSVC in two places that no cross-build can
satisfy: it asked for static Boost, which the distribution packages do not ship,
and named the Win32 system libraries with a .lib suffix, which CMake treats as a
literal file path. crypt32 and shlwapi have to be named explicitly because MinGW
does not honour the SDK headers' #pragma comment(lib).

Fedora packages mingw64 Qt 6 including Qt Multimedia, so only libvpx, Poco and
libsndfile are built from source in the image.

+++ src/SSL.cpp
@@ -123,8 +123,8 @@ void MumbleSSL::addSystemCA() {
 		sslCfg.setCaCertificates(filteredCaList);
 		QSslConfiguration::setDefaultConfiguration(sslCfg);
 
-		qWarning("SSL: CA certificate filter applied. Filtered size: %i, original size: %i", filteredCaList.size(),
-				 caList.size());
+		qWarning("SSL: CA certificate filter applied. Filtered size: %lli, original size: %lli",
+				 static_cast< long long >(filteredCaList.size()), static_cast< long long >(caList.size()));
 	}
 #endif
 }
